Sort Colors
Link:https://leetcode.com/problems/sort-colors/

3 Pointer Approach is the optimal solution for this question,instead of Counting Method.

When we get 0 we swap low with medium,then move low and medium, next location ahead(++).We are using low to sort the 0's.

When we get 1,we just move the medium since it will help in sorting 1's.

When we get 2,we swap high with medium,then decrease the high(--).
We are using low to sort the 2's.

We run this loop untill medium <= high to get the desired solution.