Duplicate Numbers in an Array
Link:https://leetcode.com/problems/find-the-duplicate-number/

1st Approach
Sort the array,and compare the ith and i+1th element,if they are equal,that means it is a duplicate element,then return the duplicate element.
Time Complexity : O(n log n)
Space Complexity : O(n)

2nd Approach
We know that element present inside array will be only from 1 to n.
We use index variable to store the value for each index,also absolute values,marking it all positive,then we mark the number present at that index as -1.
if it is already visited,we can check this condition if the value of nums[i]<0,it we will return the index,so it will give the duplicate number.

3rd Approach
Taking the values to thier index postitions.It does not modifies the array.
Time Complexity : O(n)
Space Complexity : O(1)