Find First Repeating Element
Link : https://practice.geeksforgeeks.org/problems/first-repeating-element4018/1

1st Approach
Use 2 for loops,We use the inner for loop from i+1,and check if any other element is repeating or not ,then return its index,But the Time Complectity becomes O(n^2).

2nd Approach
Using unordered map.We use hash to map the items and their number of occurences,if the hash value of the index is >1,we return the index and  get the repeating element.It is a better approach as compared to the first one as the Worst Case Time Complexity is O(n).