Moving Negative Elements to Left Side of the Array
Used Dutch National Flag Algorithm.

It is a 2 Pointer Approach,If the current element at low is negative we move the low pointer ahead,else if the current element at high is postive,we move the high pointer back,but if it does not satifsies this conditions we swap the elements at low and high,We stop when  the index of low becomes greater than high.