First and Last Occurence in an Array
Link:  https://leetcode.com/problems/find-first-and-last-position-of-element-in-sorted-array/

Used Binary Search Algorithm.

For finding the first occurence,When we find the target/key,we store the index(mid) in another variable and search on left side by doing e=m-1.

For finding the last occurence,When we find the target/key,we store the index(mid) in another variable and search on right side by doing s=m+1.

Since the function is of vector type,we use another vector to store the answer of firstOccurence and lastOccurence.