Odd Occuring Element in an Array
Find the odd occuring element in an array.
All elements occur even number of times except one.
All repeating elements occur in pair.
Pairs are not adjacent (there can not be more than 2 consective occurence of any element).

Find the element that appears odd number of times.
Array 1 1 2 2 3 3 4 4 3 600 600 4  4
Index 0 1 2 3 4 5 6 7 8 9   10  11 12

Observations based on the indexes: Answer would be present on even index and To the left side of the answer,In all the pairs,1st number of the pair will be present on even index,2nd numbert will be presen on odd index,On its right side,it will be vice-versa.