Basics of Character Array
char is a data type and array is a data structure.

We can directly take input in the character array or in a  sequence using for loop.By default null character is assigned in the array,when the string ends.It indicates the termination.This is when will directly do cin.It will stop when we assign a space or tab.

Instead of this,we can use cin.getline(ch,50),where ch is the character array and 50 is the maximum size of the array,it will also take spaces/tabs in the input.